Fix format specifiers
Patch by: thommey
Fixes: #1176

Needs to be tested and compile tested on various platforms if possible.
Uses difftime() to print time_t when possible because printing time_t portably is very hard.

* Add compiler format specifier checking where possible
* Fix various printf-style specifiers and arguments
* add format specifiers

+++ src/mod/modvals.h
@@ -118,7 +118,7 @@ typedef struct _module_entry {
 #endif /* STATIC */
   Function *funcs;
 #ifdef DEBUG_MEM
-  int mem_work;
+  unsigned long mem_work;
 #endif /* DEBUG_MEM */
 } module_entry;
